refactor: Move relevant files under infra
