+++ internal/infra/adapters/persistent_conversation_adapter.go
@@ -1,4 +1,4 @@
-package container
+package adapters
 
 import (
 	"context"
